update chat to v0.3.0

+++ src/Traits/ChatGroupConversationTrait.php
@@ -5,10 +5,17 @@ namespace Vns\Chatting\Traits;
 
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use stdClass;
+use Vns\Chatting\ChattingFacade;
+use Vns\Chatting\Helpers\ApiMessage;
 use Vns\Chatting\Models\ChatConversation;
 
 trait ChatGroupConversationTrait
 {
+    private function canEditGroup(ChatConversation $conversation)
+    {
+        return $conversation->group_admin_id === auth()->id();
+    }
+
     public function storeGroupConversation(array $userIDS, string $groupName, string $groupImage = null)
     {
         $conversation = ChatConversation::create([
@@ -28,44 +35,71 @@ trait ChatGroupConversationTrait
     {
         $conversation = ChatConversation::find($conversationID);
 
-        $data =  new stdClass;
-        $data->name = $name;
+        if ($this->canEditGroup($conversation)) {
+            $conversation->update(['name'  => $name]);
 
+            if ($image) {
+                $conversation->clearMediaCollection();
+                $conversation->saveMedia($image);
+            }
 
-        $conversation->update(['name'  => $name]);
-
-        if ($image) {
-            $conversation->clearMediaCollection();
-            $conversation->saveMedia($image);
+            return ApiMessage::success('group info updated');
         }
 
-        return $conversation;
+        return ApiMessage::error('you are not the admin');
     }
 
     public function changeGroupAdmin(int $conversationID, int $userID)
     {
         $conversation = ChatConversation::find($conversationID);
 
-        $conversation->update(['group_admin_id' => $userID]);
+        if ($this->canEditGroup($conversation)) {
+            $conversation->update(['group_admin_id' => $userID]);
 
-        return $conversation;
+            return ApiMessage::success('group admin changed');
+        }
+
+        return ApiMessage::error('you are not the admin');
     }
 
     public function addMembersToGroup(int $conversationID, array $userIDS)
     {
         $conversation = ChatConversation::find($conversationID);
 
-        $conversation->users()->syncWithoutDetaching($userIDS);
+        if ($this->canEditGroup($conversation)) {
+            $conversation->users()->syncWithoutDetaching($userIDS);
 
-        return $conversation;
+            return ApiMessage::success('members added to group');
+        }
+
+        return ApiMessage::error('you are not the admin');
     }
 
     public function deleteMemberFromGroup(int $conversationID, int $userID)
     {
         $conversation = ChatConversation::find($conversationID);
 
-        $conversation->users()->detach($userID);
+        if ($this->canEditGroup($conversation)) {
 
-        return $conversation;
+            // admin
+            if ($conversation->group_admin_id === $userID) {
+                // check if group is empty
+                if (count($conversation->users) <= 1) {
+                    $conversation->users()->detach($userID);
+
+                    $conversation->delete();
+
+                    return ApiMessage::success('member removed from group and group is deleted');
+                } else {
+                    return ApiMessage::error("admin can't left the group when the group not empty");
+                }
+            }
+
+            // normal member
+            $conversation->users()->detach($userID);
+            return ApiMessage::success('member removed from group');
+        }
+
+        return ApiMessage::error('you are not the admin');
     }
 }
